Enum rust_cast::channels::media::MediaResponse

pub enum MediaResponse {
    Status(Status),
    LoadCancelled(LoadCancelled),
    LoadFailed(LoadFailed),
    InvalidPlayerState(InvalidPlayerState),
    InvalidRequest(InvalidRequest),
    NotImplemented(String, Value),
}

Represents all currently supported incoming messages that media channel can handle.

Variants
Status(Status)

Statuses of the currently active media.

LoadCancelled(LoadCancelled)

Sent when the load request was cancelled (a second load request was received).

LoadFailed(LoadFailed)

Sent when the load request failed. The player state will be IDLE.

InvalidPlayerState(InvalidPlayerState)

Sent when the request by the sender can not be fulfilled because the player is not in a valid state. For example, if the application has not created a media element yet.

InvalidRequest(InvalidRequest)

Error indicating that request is not valid.

NotImplemented(String, Value)

Used every time when channel can't parse the message. Associated data contains type string field and raw JSON data returned from cast device.