Enum rust_cast::channels::media::StreamType

pub enum StreamType {
    None,
    Buffered,
    Live,
}

Describes the way cast device should stream content.

Variants
None

This variant allows cast device to automatically choose whatever way it's most comfortable with.

Buffered

Cast device should buffer some portion of the content and only then start streaming.

Live

Cast device should display content as soon as it gets any portion of it.