Struct rust_cast::channels::media::Media

pub struct Media {
    pub content_id: String,
    pub stream_type: StreamType,
    pub content_type: String,
    pub duration: Option<f32>,
}

This data structure describes a media stream.

Fields
content_id: String

Service-specific identifier of the content currently loaded by the media player. This is a free form string and is specific to the application. In most cases, this will be the URL to the media, but the sender can choose to pass a string that the receiver can interpret properly. Max length: 1k.

stream_type: StreamType

Describes the type of media artifact.

content_type: String

MIME content type of the media being played.

duration: Option<f32>

Duration of the currently playing stream in seconds.