Struct rust_cast::channels::media::StatusEntry

pub struct StatusEntry {
    pub media_session_id: i32,
    pub media: Option<Media>,
    pub playback_rate: f32,
    pub player_state: PlayerState,
    pub idle_reason: Option<IdleReason>,
    pub current_time: Option<f32>,
    pub supported_media_commands: u8,
}

Detailed status of the media artifact with respect to the session.

Fields
media_session_id: i32

Unique ID for the playback of this specific session. This ID is set by the receiver at LOAD and can be used to identify a specific instance of a playback. For example, two playbacks of "Wish you were here" within the same session would each have a unique mediaSessionId.

media: Option<Media>

Full description of the content that is being played back. Only be returned in a status messages if the Media has changed.

playback_rate: f32

Indicates whether the media time is progressing, and at what rate. This is independent of the player state since the media time can stop in any state. 1.0 is regular time, 0.5 is slow motion.

player_state: PlayerState

Describes the state of the player.

idle_reason: Option<IdleReason>

If the player_state is IDLE and the reason it became IDLE is known, this property is provided. If the player is IDLE because it just started, this property will not be provided. If the player is in any other state this property should not be provided.

current_time: Option<f32>

The current position of the media player since the beginning of the content, in seconds. If this a live stream content, then this field represents the time in seconds from the beginning of the event that should be known to the player.

supported_media_commands: u8

Flags describing which media commands the media player supports: * 1 Pause; * 2 Seek; * 4 Stream volume; * 8 Stream mute; * 16 Skip forward; * 32 Skip backward. Combinations are described as summations; for example, Pause+Seek+StreamVolume+Mute == 15.