Enum rust_cast::channels::receiver::ReceiverResponse

pub enum ReceiverResponse {
    Status(Status),
    LaunchError(LaunchError),
    InvalidRequest(InvalidRequest),
    NotImplemented(String, Value),
}

Represents all currently supported incoming messages that receiver channel can handle.

Variants
Status(Status)

Status of the currently active receiver.

LaunchError(LaunchError)

Error indicating that receiver failed to launch application.

InvalidRequest(InvalidRequest)

Error indicating that request is not valid.

NotImplemented(String, Value)

Used every time when channel can't parse the message. Associated data contains type string field and raw JSON data returned from cast device.