Struct rust_cast::channels::receiver::Application

pub struct Application {
    pub app_id: String,
    pub session_id: String,
    pub transport_id: String,
    pub namespaces: Vec<String>,
    pub display_name: String,
    pub status_text: String,
}

Structure that describes currently run Cast Device application.

Fields
app_id: String

The identifier of the Cast application. Not for display.

session_id: String

Session id of the currently active application.

transport_id: String

Name of the pipe to talk to the application.

namespaces: Vec<String>

A list of the namespaces supported by the receiver application.

display_name: String

The human-readable name of the Cast application, for example, "YouTube".

status_text: String

Descriptive text for the current application content, for example “My vacations”.