Module rust_cast::channels::media

Structs

GenericMediaMetadata

Generic media metadata.

Image

Image URL and optionally size metadata.

InvalidPlayerState

Describes the invalid player state error.

InvalidRequest

Describes the invalid request error.

LoadCancelled

Describes the load cancelled error.

LoadFailed

Describes the load failed error.

Media

This data structure describes a media stream.

MediaChannel

MovieMediaMetadata

Movie media metadata.

MusicTrackMediaMetadata

Music track media metadata.

PhotoMediaMetadata

Photo media metadata.

Status

Describes the current status of the media artifact with respect to the session.

StatusEntry

Detailed status of the media artifact with respect to the session.

TvShowMediaMetadata

TV show media metadata.

Enums

IdleReason

Describes possible player idle reasons.

MediaResponse

Represents all currently supported incoming messages that media channel can handle.

Metadata

Generic, movie, TV show, music track, or photo metadata.

PlayerState

Describes possible player states.

ResumeState

Describes the operation to perform with playback while seeking.

StreamType

Describes the way cast device should stream content.